Refactor GitHub Actions workflows and update configuration handling in the extension
